MacOS, Win: set font size using GtkSettings object
This commit undoes all the CSS hacks added in 04fdc77a to make the font
larger in macOS. It's much reliable to just override the default
application font in code by setting the correct properties in
GtkSettings.

Set the font for Windows larger too, but I should test later on Windows
that it looks OK.

Also, sets the monospaced font to a hardcoded one that is well known
from both operating systems, given that there is no settings schema that
can be poked to get the value.