Use native window controls and menu bar
This applies to the macOS version. This commit will add a menu bar to
Cartero, now that GTK 4.18 has better support for menu bars, including
an Edit menu. Many menu items are just delegates for existing
application and window actions, but also I've added a couple of new
actions for the Help menu.

There is a new app.about action that delegates into win.about, but this
is required because it seems that the menu bar about menu in macOS is
hardcoded to use app.about, so there has to be a handler there for it.

The native window controls are only used on macOS. They require
GtkHeaderBar rather than AdwHeaderBar, because while these were added to
GTK 4.18, the patch required to make AdwHeaderBar use native header
controls did not make the cut for libadwaita 1.7. Fortunately I am still
not using features specifically made for AdwHeaderBar, so this should
cause no harm on GNU/Linux. (Windows is not tested here because in
Windows header bars are not used at the moment).