Fix compilation issues on macOS
It seems that some dependencies are now being linked using rpaths. This
is good, but also the macos-build.sh script does not know how to
relocate object code linked with rpath, so I had to teach this to the
script.

Then, fix src/meson.build because due to the way CSD is enabled by
default to support flows using cargo run, the --no-default-features flag
has to always be passed when compiling from Meson.