Windows: show CarteroWindow when dialog closes
As reported in https://github.com/danirod/cartero/issues/128, when
closing a dialog that is not attached to a window, if any other window
was opened in the computer while the dialog was opened, the focus will
go back to the second window, not to the CarteroWindow.

This only affects Windows because it is using GtkWindow rather than
AdwWindow (something that it's currently outside of my control due to
some bugs in GTK and libadwaita), so the AdwDialog does not have a valid
widget to stick in, therefore it becomes detached.

To fix this, this commit will manually present the CarteroWindow that
originally popped the AdwDialog for Settings or About whenever the
AdwDialog closes.

This has to be tested in other OSes to make sure that there is no
regression.