Manually call win.check-updates from settings
On Windows, where the SettingsDialog is not connected to the
CarteroWindow it belongs to, because the lack of connection it is not
possible for the actions bound to buttons in the dialog to operate on
the parent window.

Therefore, I am replacing the action in the Check updates button for the
settings dialog with a callback and event that just broadcasts the
action manually, for compatibility with the different OSes.