Internationalize response units
There is a function called glib::format_size() to format a numerical
amount of bytes into a humanized string such as 431 bytes, 23.1 kB or
5.4 GB. The cool thing is that this one has internationalization, so
running the application in French will render the units as 431 octets
and 23.1 Mo.

The duration is now internationalized in cases where it makes sense. I'm
reducing the precission of the seconds to the hundreth part of a second
to keep it simple.