Create GObject models
Rebasing to the split previously made to separate each entity into a
separate submodule, it is more easy to see now each entity and thus it
is easy to convert every entity into a GObject.

This commit adds a crate to store the modules and converts Cartero into
a workspace. I'm not a fan of monorepos, but this allows to expose the
public module API without the inners, and to make sure that no GTK code
is accidentally added to the object modules, because cartero_objects
only depends on GLib and GIO.

This commit is the point of the whole refactor, but adding more tests
now that the models are separated would be nice to make sure that it
works fine. Some of these tests could be written as I change the UI so
that it uses the crate instead of the entities module.