[BACKPORT] Use a lax parser for queryparams in UI
In this commit, the callbacks that run after updating the URL bar or the
query table to update the other component are changed so that instead of
parsing the URL using the real url crate to get the query params, these
are extracted from the raw string.

- To extract the query parameters from the URL in the address bar, the
  new extract_queryparams function is used. Because the "?" symbol is
  not valid neither for the scheme, the host or the path (according to
  the URL spec), any ? in the URL must separate the querystring, so can
  be trimmed. Then it's just a matter of decoding the querystring as a
  urlencoded string.

- To encode the params as a string, the new update_queryparams function
  is used. This function removes the querystring from the existing URL,
  then serializes the table using the algorithm described in the spec,
  and sets the URL address to that specific URL.

These functions DO NOT actually validate an URL. However, because the
user is still typing an URL, it doesn't matter. The URL gets validated
before the request. The URL cannot be validated when these pair of
functions are called because the user may still be typing the characters
of the URL, so of course it's not valid yet.