Ignore unbound variables in disabled headers
If a header is not active, trying to render the value of the header and
thus potentially falling into SrTemplate errors if there is an undefined
is not worth.

Actually, not a big deal but why rendering headers that are not going
to be used at all and that will be ignored in the actual HTTP request.

This commit therefore changes the render() function to prevent it from
rendering disabled variables.