[BACKPORT] Make the query params table less laggy (#204)
It seems the reason why the URL field and the tables feel so laggy
sometimes is because destroying and re-creating every row every time
KeyValuePane.set_model() is called is not efficient (who could guess).

Therefore, the more queryparams your URL has, the more rows has to
destroy and then recreate, which causes the textfield used to set the
URL to behave more laggy, as it has to wait until the table updates to
finish the event handler.

To fix this, the setter for KeyValuePane:model will not destroy the
ListStore and create it again. Instead, if possible, it will try to
update the values of existing rows in the ListStore, and then only
create new rows if the size has grown, or delete the remaining rows if
the size has shrunk.

I've added tests to check that the behaviour of set_model() is
consistent and doesn't lose information. This includes some UI tests at
EndpointPane that assert the connection between the address bar and the
parameters table.