Allow the user to configure the app language
Adds a new schema key called `locale`. The value can be either an empty
string to use the system default, or one of the ISO codes in po/LINGUAS,
to force Cartero to use that language.

Then adds a configuration option to the settings page in order to
present the list of LocaleRepr (the object that will represent a locale
and its ISO code), which is bound to the settings in order to save the
value.