Dynamically use the list of languages
In this commit, the list of languages presented in the settings will
only contain locales for which there is actually a .mo file in the
localedir.

The reason behind this change is that some platforms such as Flatpak
actually split the locales package and only download locales for the
languages configured in the computer to save space. Therefore, in
Flatpak most of the time the user is not downloading all the languages.