Fix deadlock issues over static_type
Apparently running a subset of the current test suite causes a thread
lock if multiple tests run in parallel (the default behaviour of
`cargo test` unless the number of threads is changed).

It turns out that the tests manage to register some types at the same
time, causing a deadlock because they acquire the same futex. This thing
is nuts and debugging it has been difficult, but if it happens, it can
be fixed by making sure that the static_type() is called in a safe
location to register it globally before starting the tests.

This commit adds ctor as a way to fix this. Currently only
FieldTable::static_type() is called, but if more tests start to fail,
more calls could be added or, in fact, ctor could be added as a
workspace dev-dependency.