Fix prefix oddities on MSYS environment
Unlike a shell script, calling Meson right from an MSYS shell will cause
Meson to use Windows paths (C:\msys64\ucrt64) rather than POSIX paths
(/ucrt64). Therefore, the prefix argument shall be given Windows style,
or else additional prefix paths will be present in DESTDIR. This commit
makes it use a Windows slash (\) as a prefix, which will cause the
DESTDIR to be the actual root directory as long as the same disk is
used (which for Windows after reading the GitHub Workflows log it seems
it is because it is D:\).