Remove unnecessary runners for macOS
macos-14 is totally not needed.

I'm keeping macos-13 because it is the x64 runner (even though I suppose
it will be retired after summer when macOS 26 is released, due to the
Homebrew policy), and macos-15 because it is the arm64 runner. I suppose
that after summer, the macos-13 runner will be removed the moment it
starts to fail.