[BACKPORT] Rename appdata to metainfo
Apparently using /usr/share/appdata/*.appdata.xml is not deprecated but
outdated, and appstream files should be placed as
/usr/share/metainfo/*.metainfo.xml now.

In this commit, the installation target for the appstream file is
changed to metainfo. It is also renamed, therefore a lot of files have
to be touched as well. The POTFILES.in has the updated location for the
filename, and thus every translation has updates.