Integrated packaging scripts and ARM64 support (#218)
This change does three things:

* Merges the existing packaging scripts into the Meson infrastructure.
Extra options are added to the meson_options.txt file and now running
the `install` target will allow the binary distribution to be bundled
and signed.
* Rewrites the packaging scripts in Python. For the kind of things they
do, it is more convenient to use `shutils`, `subprocess` and `pathlib`
and have features of a high level language rather than Bash scripting.
* Add CI pipelines to build ARM versions of Windows, AppImage and
Flatpak for the very first time using native Windows 11 ARM and Ubuntu
aarch64 runners. The AppImage aarch64 will be released as an official
artifact for the first time in future releases.

## Visible changes for users (this goes in the changelog)

**Cartero's AppImage will now work on legacy and oldstable GNU/Linux
distros**, where previously an error message about the glibc version in
the system would be presented. We're packaging an updated version of
glibc in the AppImage now, allowing the application to work on most if
not all oldstable GNU/Linux distros.

This change has required to make tweaks to the build pipeline. **If you
are finding issues after upgrading to Cartero 0.2.3 with the AppImage
version, or if the AppImage has stopped working for you, please open an
issue**.

For the first time, **Cartero is available as an AppImage for the
aarch64 architecture**. If you are doing web development on an ARM
computer, like a Raspberry Pi, an ODROID, a Pinebook, Pinephone or
Pinetab, or even an Android device running a graphical GNU/Linux
environment, you can use the AppImage executable to quickly start
testing your APIs in a lightweight graphical environment.

Also, **we have recently checked that Cartero can compile for the
Windows 11 ARM platform**. The good news is that Cartero compiles
successfully. The bad news is that I have no way to run Windows 11 ARM,
so I cannot check that the program actually opens or if it will display
an error message on launch. Therefore, there is no download link at the
moment. **If you have a computer running Windows 11 ARM and you want to
help testing, get in touch! You could help literally make history**.

## Visible changes for developers and packagers

Release engineering process has been simplified by replacing the
existing packaging and vendoring scripts for Windows, macOS and AppImage
with new build scripts that are called by Meson when calling the
`install` target if the meson options are properly set.

Check out the release engineering docs to see the full commands and
requirements, but with the dependencies installed, it is now possible to
run `DESTDIR=$somewhere ninja -C build install` to generate a valid
Windows version either as .zip or as an installer, as well as MacOS .app
and .dmg bundles and Linux AppImage files.