Fix NixOS flake (#219)
After making meson.build depend on the build-aux/gen-version.py script,
it is now required to have Python 3 available to compile Cartero.

Patching the build-aux/gen-version.py shebang is now required because it
needs to point to the real python3.

This branch will also add GitHub Actions integrations to check that the
flake.nix works correctly.