Deploy to Snapcraft-Edge every night (#220)
This change will add a GitHub Action that triggers every night. If there
is any new commit in the repository, it will compile an updated
Snapcraft package based on the snapcraft.yaml recipe and upload it to
the latest/edge channel of the [Snapcraft
package](https://snapcraft.io/cartero).

The recipe is marked with grade=devel to prevent accidentally releasing
it to stable, because it is a development build that contains code that
may not be finished.