Move file serialization to a separate crate (#191)
Moves all the serde entities from src/file.rs to a new crate called
**cartero-file-format**. This crate already uses the GObject models
added in #137, and it is a required step in order to delete the usage of
the cartero::entities mod. Some advantages of the new crate are:

* It only depends on GLib and Gio, but not in GTK, so it is easier to
write tests. (And boy I am testing the heck of it).
* It allows to declare some structs as `pub(crate)` knowing that they
are still not visible from outside the crate, so it is easier to split
the structs in multiple files.
* Real integration tests. The intention is to have a folder full of TOML
files with a suite of tests that asserts that every one of them works as
expected. (Currently these tests are written as unit tests in
src/file.rs, but this requires some hacky functions declared as
`#[cfg(test)]` in order to expose the inner structures.

Some common data structures such as the warning tags are being moved to
a second new crate called **cartero-interop**. The intention is to have
some common layer for things that could be used by other file loaders
and savers. In fact, once this PR is merged, this will unlock #97, which
will be implemented as a separate crate that is easier to test and
develop without requiring a whole UI. That would also allow testing
whether cartero-interop is good enough.