Define cartero_http crate
cartero_http is another workspace crate. It provides the same
functionality that the cartero::client mod currently does in order to
provide a BoundRequest object, but this one is backed by the objects
defined in cartero_objects.

Also, this has been tested a lot.

Note that this commit will also add additional methods to deal with
FieldTable rendering (such as materializing variables, filtering active
items...) to cartero_objects. Those calls have been added to
cartero_objects because if the refactor continues, I feel I'm going to
use them for more parts of the application.