Replace ResponseData with equivalent object
cargo::entities::ResponseData is gone and usages of the struct are
replaced with interactions against cartero_objects::Response. The code
compiles, the entities are still assigned as expected to the user
interface. As a side effect, you can now browse request and response in
the GTK inspector, so this looks very promising.