Make the ResponseHeaders widget reactive
The ResponseHeaders widget previously was connected to a ListModel
instance that was created as soon as a new response was received.
Because the Response is now object oriented, the ResponseHeaders now has
the FieldTable itself as a property.

A new widget called FieldActionRow has been added. It is a
specialization of the Adw.ActionRow, but it already has a Field as a
property, and the widget setups binds so that it always reflect the key
and the value of the inner Field.

Through the magic of bindings, the ResponseHeaders ListBox uses the
FieldTable (upcasted to a ListModel) to automatically bind every Field
in the ListModel into a FieldActionRow. This is reactive and changing
the headers of a Response object will now automatically update the UI.