Scaffold a new menu button for exporting things
Additional actions are added to window, which currently do not do
anything but will delegate the event into the EndpointPane.

The actions have been configured so that they require an EndpointPane to
be usable, and some of them additionally require the EndpointPane to
have successfully made an HTTP request to be usable (exporting response
data requires response data in first place).