Implement the new export dialog
This dialog receives the blob and the file type before being presented.
When presented, it will display if possible the contents of the blob as
a chararray, and allow to either copy to the clipboard or save as a
file.

In this commit, all the features currently provided by the export pane
are present in the new export dialog. Pressing the "copy to clipboard"
button will have the same effect as pressing the old "copy" button in
the export pane.

Note that currently the "Export as HTTPie" and "Export as JetBrains
HTTP" do exactly the same as the "Export as cURL" button, because these
exporters are not implemented at the moment.