Quick reconcile for FieldTable
This is a bypass to update the contents of a FieldTable reusing existing
Field instances, and GTK widgets if the FieldTable is being mapped into
a GtkListBox, GtkListView or any widget that could be potentially bound
to a GioListModel.

The main use case of this function will be updating the Query parameters
pane whenever the entry URL changes. Currently this is done using a set
of functions in EndpointPane, but as part of the refactor to use
GObject, new objects have to be used as targets.