Assert code is compliant with the REUSE specification (#234)
[The REUSE specification](https://reuse.software/) provides tools and
procedures to make licensing easy.

To make the project compliant with the spec, every file has to be
attached with some licensing metadata that describes which license terms
has each file. The metadata can either be in the file itself, usually at
the header, or in a sidecar file such as a .license file or a REUSE.toml
file that describes in bulk many files.

## Assessment of the licenses use in Cartero

For files part of the project:

* Cartero is licensed as GPL-3.0-or-later. This includes the source
code, some metafiles in the data/ directory and the translations. Build
scripts like the Cargo.* files or the *.nix files are also affected by
this.
* The Cartero icon and logo is CC-BY-SA-4.0.
* Some files are either generated or are just config files or demo
files, so they lack any creative interest. This includes the .cartero
test files. These are CC0-1.0. The metainfo.xml is also CC0-1.0, because
the file was already licensed like this.
* The documentation and README files are CC-BY-SA-4.0.
* The website is GPL-3.0-or-later OR CC-BY-SA-4.0, depending on whether
you look at the code or at the text.
* The icons designed for social media (avatar, profile picture, profile
banner, screenshots) are released as CC0-1.0 anyway, because it's
difficult to control the distribution for them anyway.

For reused files:

* Some files are taken from the
[gtk-rust-template](https://gitlab.gnome.org/World/Rust/gtk-rust-template)
project, which is MIT.
* Some files are taken or derived from the [Adwaita Icon
Theme](https://gitlab.gnome.org/GNOME/adwaita-icon-theme), which are
CC-BY-SA-4.0.
* Some files are part of the Adwaita iconset provided by the [Icon
Library app](https://gitlab.gnome.org/World/design/icon-library), which
are CC0-1.0.
* Some files for vendor logos and such used in the website download page
are taken from their respective source pages at Wikimedia Commons,
either under CC0-1.0, CC-BY-3.0, CC-BY-4.0 or GPL-3.0-or-later license,
depending on the asset.

## Compliance

Some files lacked the license identifiers. For source code files or
where a comment can be added at the top of the file, the metadata was
provided via comments. For those that are binary or external, they were
added to a REUSE.toml file.

This commit will also delete some third party assets such as Adwaita
icons that were never used in the final version of the website. If
they're ever needed for a new design, they can be downloaded again
anyway.