objects: add data structures for RequestBodyFile
RequestBodyFile is a new kind of RequestBodyData that indirectly points to
a file with the payload. When the HTTP request is submitted, the pointed
file is read, and the request body of the HTTP request is assigned to the
contents of the file.

This is NOT attaching files to a multipart request. This is more like a
RequestBodyRaw request, except that the payload is read from a file when
the request is sent, rather than being manually copied into the data
structure.

At a file level, the request file only stores the path to the file whose
contents have to be read when the request is sent, but the contents of the
file are not actually stored in the request itself.

Optionally, it is also possible to set a content type for the given file.
The UI may try to guess the correct MIME type for the given file when
picked from the user interface and pre-populate this field to an acceptable
value. For instance, if the user picks a .png file, the content type may be
automatically set to image/png. The default value will be
application/octet-stream.

In this commit, the object RequestBodyFile is created with the following
properties:

- path (string): the path to the file to be used as payload.
- content_type (string, nullable): the content type that will be
  automatically asigned to the Content-Type header of the given request.

Because this commit introduces new items to the RequestBodyType enum, the
code for the file serialization and deserialization is also provided in
this commit; and so is a placeholder for the actual code of the HTTP
client, which will be added later.

ChangeLog items:

* cartero-objects/src/request_body.rs: A new value has been added to
  RequestBodyFile: 'RequestBodyFile::File'.