ui: add a relative file picker
cartero::widgets::file_dialogs::pick_file() is a function that presents a
GtkFileDialog and lets the user pick a file from the file system in 'read'
mode. There are no file filters and any file can be picked, no matter the
extension.

However, the function receives as a parameter a GioFile to be used as
prefix of the section. This prefix must be a directory. After the file has
been selected by the user, it is validated. The selected file must either
be in the prefix directory, or in a subdirectory inside the prefix
directory. It is not possible to pick absolute paths or files that are
outside of the prefix directory.

This is done by design because of three reasons:

1. It reduces the surface of an attack where a malicious user fools an user
   into sending a random file from the hard drive such as /etc/password or
   ~/.npmrc as an HTTP request.

2. It makes the request and collection files more portable when there are
   multiple users working with the files (in a Git repository, for
   instance); because users may clone the repository in a different
   location of the file system, but the relative path would be the same, as
   long as the fixtures and the request files are part of the file system.

3. It is easier to undo this policy if I'm convinced that it's actually
   safe and/or convenient to allow picking any file from the file system.
   Limitting the access if you could pick any file in first place may be
   more difficult because some existing requests could break.