ui: add widget to drive file payloads
CarteroFileBodyPane is a pane that allows setting the options for the path
and the content_type properties of a CarteroRequestBodyFile object.

The user can manually type the path and the content_type because the
intention is that the path can have variables to expand (although this may
be a security concern), and the content_type can be set to whatever the
user wants.

However, the pane also has a button for quickly picking a file from the
file system and pre-filling the values with the relative file path and the
guessed content type.

This can only be done if the request file has been saved first, because
otherwise it is impossible to guess the prefix directory to be used for
relative paths. The button is disabled if the request is anonymous.