http: refactor BoundRequest instantiation
BoundRequest will now require the RequestEnvironment to initialise. Some
aspects of the request conversion, such as the current project root, will
come via the environment until a better solution is found out.

Also, because dealing with files is an I/O operation, instantiating a
BoundRequest now is an async process.

Overall, this commit replaces the old BoundRequest::try_from() function by
an async function called new() that takes two parameters: the Request to
bind, and the RequestEnvironment to use.

Changelog:

* cartero-http/src/request.rs: BoundRequest now requires the
  RequestEnvironment in order to be instantiated. Therefore,
  BoundRequest::try_from() has been replaced by BoundRequest::new(), so
  that the function can accept two arguments.
* cartero-http/src/request.rs: BoundRequest is now async.