ui: don't render binary responses by default
When an HTTP response succeeds and returns a body that contains
non-printable characters, or when it uses a content type such as image/,
audio/ or video/, then the response should be treated as binary by
default.

Binary responses will not present the CodeView widget by default.
Instead, this commit will add a new panel that only appear when a
response has binary data, with two buttons:

* Export the response body into a file (using the recently added
  exporter).
* Force rendering the response as text in the CodeView anyway.