Defer assignment of response payload into buffer
With the previous implementation of the new binary response detection,
the contents of the response are assigned to the CodeView TextBuffer
anyway, even if the response may never be rendered.

This commit will only assign the response payload to the CodeView if the
response is text, or if the user has forced the response to be rendered
as text.