ui: only show share menu when an endpoint is open
In the future, there will be additional types of pane being added to the
CarteroWindow TabView. Therefore, some widgets in the toolbar will only
make sense if the current pane is of a specific kind.

This commit will make the share menu only be visible when the pane is an
EndpointPane. Is the only type right now, but soon it will not be. Also,
this commit will enable the next refactor: the export request and
response actions should actually be part of the EndpointPane, and the
window actions should just defer to the appropiate action in the current
pane.