Release to Snapcraft through GitHub too
The build service offered by Snapcraft is too limiting IMHO. For
instance, it always overrides the nightly branch with the latest stable
after release, and I don't want that behaviour.

Also, I cannot release to multiple channels at the same time.

This commit will add Snapcraft recipes both for stable and for the
beta branch. They are based in danirod/cartero-snap, which will be
retired when this works properly.