Keep and restore old auth and body data in widget
With this change, switching the auth or body type will keep in memory
the old authentication or body data that is just being replaced. This
allows to put the data back in if the old type is picked again, rather
than starting with an empty form.

This will fix https://github.com/danirod/cartero/issues/118 at the same
time because now we keep and restore the parameter table, so it can be
restored when the method switches between multipart and urlencoded.