Implement Request::resolve()
This method will return a copy of the Request object, but the variables
have been already replaced in the request. In other words, if you call
resolve() in a request that points to an URL like "{{API_ROOT}}/users",
and you have a variable called "API_ROOT" with a value, calling this
method will resolve the value and render the templates.