Use specific widgets for the static tables
Because the user will not be able to touch things in the pre-generated
headers, I'd rather hide the checkbox and the menu button. I suppose
there is probably a better way to do this with a single ListBox if I
have a way to differentiate pre-generated and non-pre-generated headers
in the same List and I map the entries to a different widget depending
on the area they belong to, but this is a good start anyway.