Use the new crate for the export dialogs
This commit replaces the old modules used by the export system with calls
to the functions provided by the code_exporters crate. These are called by
the EndpointPane export function. Because the new crate provides its own
error type, mappers to an error message are provided.

Also, the cartero-jetbrains-http-format crate is deleted because it won't
be used anymore, and the old export module to export as curl is removed too
because it is not being used anymore.