Remove existing code for app callbacks
The purpose of this commit is to remove the code from the activate() and
open() functions, as well as scrap the setup_app_actions().

The new shell uses completely different code. This commit exists to help
dealing with the possible conflicts when the branch is rebased on top of
the current trunk.