Upgrade GNOME SDK to 46 and rest of Rust deps (#270)
glib 2.80, gtksourceview 5.12 and gtk 4.14. This is required as part of
#238 and GtkSourceView 5.12 will actually be required to check for
themes in order to implement the theme changer.

gtk-rs now actually checks that the widgets implement the required
interfaces, so if you have a widget that extends gtk::Box, it is now
required to implement gtk::Accessible, gtk::Buildable and
gtk::ContsraintTarget and so on. That's why this dependency bump also
has a lot of changes to the widgets.

Finally, I am updating the rest of the dependencies because why not.