Provide clipboard variants for macOS
macOS cannot accept the clipboard contents unless it has the charset
attribute set. Therefore, add a variant that specifically sets the
chaset when copying the contents of an export.

The text/plain is kept because it works on Linux and probably Windows
(need to test), and I don't want to break stuff.

Fixes https://github.com/danirod/cartero/issues/275