win32: fix custom language not being applied (#280)
It seems that on Microsoft Windows I cannot just set the LANGUAGE env
var during the initialisation phase to override the application language
with the one read from the GSettings schema.

Therefore, on Windows I'll just restart the application with the new
environment variable value if it has to change. This has the consequence
that overriding the language will actually spawn two cartero.exe
processes.