Win32: Use acrylic design
This commit will switch to an acrylic style when the application runs
on Windows and the CSD is disabled (which should be the default,
anyway).

In acrylic mode, the toolbar is transparent and it merges with the title
bar. Some elements of the user interface such as window backgrounds are
also slightly acrylic. The view used for EndpointPane stays solid.