Render response headers as raw
It seems that the `use-markup` property of the Adw.PreferencesRow is
enabled by default. This causes issues when the header is a Location
header with an ampersand in it, because it treats the ampersand as an
HTML entity that is unable to escape.

For safety, these should be treated raw and not rendered as HTML by
default, to prevent possible security issues.

Fixes #291