Remember the visibility status of the tokens
By default, the basic auth token will be hidden as a password field, and
the basic auth password too. This was the behaviour in 0.2.4 and earlier
versions.

However, toggling the visibility of the password will now be remembered,
and it will be saved as a setting, allowing the token and password to be
visible if the user presses the button.