Menu to duplicate a request
The duplicate request menu creates a copy of the current endpoint and
opens it as a new pane. The copy of the endpoint is made using .dup(),
so it is completely different and independent, and changes do not affect
the original endpoint. The new endopint is treated as a new endpoint and
has to be saved.